refactor: restructure fixtures and ignore some deprecated message
