+++ tests/integrations/example/api/v1/sys/test_health.py
@@ -1,6 +1,7 @@
 """Test cases for the health endpoint."""
 
 import os
+import asyncio
 from http import HTTPStatus
 from typing import Any
 from unittest.mock import patch
@@ -8,10 +9,10 @@ from unittest.mock import patch
 import pytest
 from fastapi.testclient import TestClient
 from httpx import Response
-from motor.motor_asyncio import AsyncIOMotorDatabase
 from structlog.stdlib import get_logger
+from testcontainers.mongodb import MongoDbContainer
 
-from fastapi_factory_utilities.example import AppBuilder
+from fastapi_factory_utilities.example.app import AppBuilder
 
 _logger = get_logger(__package__)
 
@@ -20,16 +21,12 @@ class TestApiV1SysHealth:
     """Test the health endpoint."""
 
     @pytest.mark.asyncio()
-    async def test_get_api_v1_sys_health(self, async_motor_database: AsyncIOMotorDatabase[Any]) -> None:
+    async def test_get_api_v1_sys_health(self, mongodb_server_as_container: MongoDbContainer) -> None:
         """Test the get_api_v1_sys_health function."""
-        with patch.dict(
-            os.environ,
-            {
-                "MONGO_URI": async_motor_database.client.delegate._init_kwargs[  # pylint: disable=protected-access
-                    "host"
-                ],
-            },
-        ):
+        # Use the container's connection URL directly
+        mongo_uri: str = mongodb_server_as_container.get_connection_url()
+
+        with patch.dict(os.environ, {"MONGO_URI": mongo_uri}):
             with TestClient(app=AppBuilder().build()) as client:
                 response: Response = client.get(url="/api/v1/sys/health")
                 assert response.status_code == HTTPStatus.OK.value
